fix(gates): split the LLVM linkage helpers out of module.rs (file cap, grown by #9610)
